build(nix-shell): use musl libc and static build
The non-buildx Dockerfile copies a binary built outside of the docker
build, which may be broken if the dependencies are not met.
This happens on nix-shell as the deps exist on /nix/store.
Get around this by building static binaries in the nix-shell.
We now depend on golang-alpine image and install musl-dev as well.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>